climate-chatbot/requirements.txt
streamlit==1.45.1
python-dotenv==1.1.0
google-generativeai==0.8.5